requirements/base.txt
# Core dependencies
Django==5.2.12
mysqlclient>=2.2.0
python-decouple==3.8

# Admin theme
django-unfold==0.82.0

# Tailwind CSS pipeline
django-tailwind==4.4.2
pytailwindcss>=0.3.0

# Async task queue (stubs — activate when implementing AI/Contests)
celery>=5.3
redis>=5.0

# API (stub — activate when implementing REST API)
djangorestframework>=3.15

# Audit trail
django-simple-history>=3.5

# Import / Export (unfold integration)
django-import-export>=4.0
